feat: implement avatar handling with sharp and refactor signup logic to use storeAvatar
