+++ bruno/API/OAuth2/Google/Sign up.bru
@@ -12,6 +12,6 @@ post {
 
 body:json {
   {
-    "username": "dorian.eydoux"
+    "username": "dorian_eydoux"
   }
 }
