+++ src/server/routes/api/account/avatar/put.ts
@@ -1,12 +1,6 @@
-import * as sharp from 'sharp';
 import {FastifyPluginAsync} from 'fastify';
-import SQL from 'sql-template-strings';
-import {join} from 'node:path';
-import {rename} from 'node:fs/promises';
 
 const plugin: FastifyPluginAsync = async server => {
-  let it = 0;
-
   await server.register(import('@fastify/multipart'), {
     limits: {
       files: 1,
@@ -28,33 +22,9 @@ const plugin: FastifyPluginAsync = async server => {
         message: 'File too large (max 5MB)',
       };
 
-    const buffer = await data.toBuffer();
-
-    let avatar;
-    try {
-      avatar = sharp(buffer);
-    } catch (error) {
-      throw server.httpErrors.badRequest(
-        error instanceof Error ? error.message : String(error),
-      );
-    }
-
-    const {height, width} = await avatar.metadata().catch(error => {
-      throw server.httpErrors.badRequest(error.message);
-    });
-    const size = Math.min(height ?? 0, width ?? 0, 1024);
-    avatar.resize(size, size);
-
     const {id} = request.user;
-
-    const cacheFile = join(server.paths.cache, `avatar_${++it}.webp`);
-    const avatarFile = join(server.paths.avatars, `${id}.webp`);
-    await avatar.toFile(cacheFile);
-    await rename(cacheFile, avatarFile);
-
-    await server.db.run(
-      SQL`UPDATE users SET has_avatar = TRUE, avatar_version = avatar_version + 1 WHERE id = ${id}`,
-    );
+    const buffer = await data.toBuffer();
+    await server.storeAvatar(id, buffer);
 
     return reply.code(204).send();
   });
